requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
allensdk==2.16.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
argschema==3.0.4
async-timeout==5.0.1
attrs==25.4.0
blosc2==2.7.1
boto3==1.42.29
botocore==1.42.29
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
einops==0.6.1
filelock==3.24.2
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
Glymur==0.14.0.post1
greenlet==3.3.0
h5py==3.15.1
hdmf==4.2.0
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.3
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
ndindex==1.10.1
ndx-events==0.4.0
nest-asyncio==1.6.0
networkx==3.4.2
numexpr==2.14.1
numpy==1.23.5
omegaconf==2.3.0
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079158/work
pandas==1.5.3
patsy==1.0.2
pillow==12.1.0
propcache==0.4.1
psycopg2-binary==2.9.11
py-cpuinfo==9.0.0
pyarrow==23.0.1
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynrrd==1.1.3
pynwb==2.8.3
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
s3transfer==0.16.0
scikit-build==0.18.1
scikit-image==0.24.0
scikit-learn==1.7.2
scipy==1.10.1
seaborn==0.13.2
semver==3.0.4
simpleitk==2.5.3
simplejson==3.20.2
six==1.17.0
snntorch==0.9.4
SQLAlchemy==2.0.45
statsmodels==0.14.6
sympy==1.14.0
tables==3.10.1
temporaldata==0.1.3
threadpoolctl==3.6.0
tifffile==2025.5.10
tomli==2.4.0
torch==2.10.0
torch_brain==0.1.0
torchaudio==2.10.0
torchmetrics==1.8.2
torchtyping==0.1.5
torchvision==0.25.0
tqdm==4.67.1
typeguard==2.13.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
xarray==2023.1.0
yarl==1.22.0
